+++ lib/enums/UrlEnum.ts
@@ -16,12 +16,9 @@
 export enum UrlEnum {
   BASE_URL = 'dev.visualwebsiteoptimizer.com',
   SETTINGS_URL = '/server-side/settings',
-  WEBHOOK_SETTINGS_URL = '/server-side/pull',
-  TRACK_USER = '/server-side/track-user',
-  TRACK_GOAL = '/server-side/track-goal',
-  PUSH = '/server-side/push',
-  BATCH_EVENTS = '/server-side/batch-events',
+  // WEBHOOK_SETTINGS_URL = '/server-side/pull',
+  // BATCH_EVENTS = '/server-side/batch-events',
   EVENTS = '/events/t',
-  ATTRIBUTE_CHECK = '/attributes/check',
-  GET_USER_DATA = '/getUserData',
+  ATTRIBUTE_CHECK = '/check-attribute',
+  GET_USER_DATA = '/get-user-details',
 }